+++ Vault/src/utopia/vault/nosql/factory/FromResultFactory.scala
@@ -78,7 +78,7 @@ trait FromResultFactory[+A]
 	  * @param connection DB Connection (implicit)
 	  * @return All items with specified row ids
 	  */
-	def withIds(ids: IterableOnce[Value], order: Option[OrderBy] = None)(implicit connection: Connection) =
+	def withIds(ids: Iterable[Value], order: Option[OrderBy] = None)(implicit connection: Connection) =
 	{
 		table.primaryColumn match
 		{